Fix: Correct __repr__ formatting for zero-value and non-standard-scale units, fixes #77
